Develop

+++ shop-database/src/main/resources/db.changelog/v-1.0/01-create-product-tables.sql
@@ -1,10 +1,10 @@
 create table products
 (
-    id          bigint       not null auto_increment,
+    id          int       not null auto_increment,
     title       varchar(255) not null,
     price       decimal,
-    brand_id    bigint       not null,
-    category_id bigint       not null,
+    brand_id    int       not null,
+    category_id int       not null,
     primary key (id)
 ) engine=InnoDB;
 GO
